requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.4
Django==4.1.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
pytz==2023.3
rest-framework-simplejwt==0.0.2
setuptools==69.0.3
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
